pyproject.toml
[project]
name = "snorkelforecast"
version = "0.1.0"
description = "Carboneras snorkelling forecast using Open‑Meteo"
requires-python = ">=3.12"
dependencies = [
    "Django>=5.0.3",
    "httpx>=0.27.0",
    "python-dateutil>=2.9.0",
    "whitenoise>=6.6.0",
    "gunicorn>=22.0.0",
    "pillow>=11.3.0",
    "pyrefly>=0.28.1",
    "python-slugify>=8.0.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.4.0",
  "pre-commit>=3.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyrefly.errors]
bad-override = false

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]
